Updated routing to the new blogs page
