+++ app/blogs/[slug]/page.tsx
