+++ app/blogs/page.tsx
@@ -1,13 +1,14 @@
-import React from "react";
-import MainSectionBlogs from "@/components/Blogs/MainSectionBlogs";
-import NavbarGroup from "@/components/Global/NavbarGroup";
-const page = () => {
+'use client';
+
+import React from 'react';
+import NavbarGroup from '@/components/Global/NavbarGroup';
+import MainSectionBlogsWP from '@/components/Blogs/MainSectionBlogsWP';
+
+export default function Page() {
   return (
-    <main className="bg-black">
-      <NavbarGroup/>
-      <MainSectionBlogs />
+    <main className='bg-black'>
+      <NavbarGroup />
+      <MainSectionBlogsWP />
     </main>
   );
-};
-
-export default page;
+}
