+++ components/Blogs/BlogPostsWP.tsx
@@ -31,7 +31,7 @@ export default function BlogCardWP(props: { post: WPPost }) {
         </div>
         {/*Right Div*/}
         <div className='w-full md:w-[40%] flex flex-col md:ml-14 my-4 mr-4 space-y-4'>
-          <Link href={`/newBlogs/${post.slug}`}>
+          <Link href={`/blogs/${post.slug}`}>
             <h2 className='text-xl md:text-3xl  text-white font-bold  cursor-pointer'>
               {post.title.rendered.replace(/&nbsp;/g, ' ')}
             </h2>
